+++ R/select.R
@@ -114,11 +114,13 @@ select <- function(data, selection, cols = TRUE) {
   # Define an error message function if wrong rows or columns are selected
   errMsg <- function(nt) {
 
-    paste(ifelse(cols, "Columns '", "Rows '"),
+    paste(ifelse(cols, "Column", "Row"),
+          ifelse(length(nt) > 1, "s '", " '"),
           ifelse(length(nt) <= 5,
-                 paste(paste(nt, collapse = "', "), "' ", sep = ""),
+                 paste(paste(nt, collapse = "', '"), "' ", sep = ""),
                  paste(paste(nt[1:5], collapse = "', '"), "', and others ", sep = "")),
-          "are not present in 'data'.", sep = "")
+          ifelse(length(nt) > 1, "are", "is"),
+          " not present in 'data'", sep = "")
 
   } # errMsg
 
@@ -129,9 +131,10 @@ select <- function(data, selection, cols = TRUE) {
     if (cols) {
 
       # Verify column names match
-      if (length(notThere <- setdiff(selection, colnames(data))))
+      if (length(notThere <- setdiff(selection, colnames(data)))) {
         stop(errMsg(notThere))
-
+      }
+      
       # Get the column indexes. This complicated code preserves the ordering of 'selection'
       cd <- colnames(data)
       selection <- unlist(lapply(selection, function(x) which(cd %in% x)))
@@ -142,8 +145,9 @@ select <- function(data, selection, cols = TRUE) {
     else {
 
       # Verify row names match
-      if (length(notThere <- setdiff(selection, rownames(data))))
+      if (length(notThere <- setdiff(selection, rownames(data)))) {
         stop(errMsg(notThere))
+      }
 
       # Get the row indexes
       rd <- rownames(data)
@@ -155,26 +159,37 @@ select <- function(data, selection, cols = TRUE) {
   # Verify that the numeric indexes are in the range of 'data'''
   else {
     if (cols) {
-      if (length(notThere <- setdiff(selection, 1:NCOL(data))))
+      if (length(notThere <- setdiff(selection, 1:NCOL(data)))) {
         stop(errMsg(notThere))
+      }
     }
     else {
-      if (length(notThere <- setdiff(selection, 1:NROW(data))))
+      if (length(notThere <- setdiff(selection, 1:NROW(data)))) {
         stop(errMsg(notThere))
+      }
     }
   }
 
-  # If the length of selection is 2 or more:
-  if (length(selection) > 1) {
-
-     if (cols)
+  # Variables required to determine the case we're in
+  nRow <- nrow(data)
+  nCol <- ncol(data)
+  lSel <- length(selection)
+  
+  # Case 1:  n x m, select > 1 row or select > 1 column
+  if ((nRow > 1) & (nCol > 1) & (lSel > 1)) {
+      
+     if (cols) {
        out <- data[,selection]
-     else
+     }
+     else {
        out <- data[selection,]
-  }
+     }
+     
+  } # Case 1
 
-  # Selecting a single row or column
-  else {
+
+  # Case 2:  n x m, select 1 row or 1 column
+  else if ((nRow > 1) & (nCol > 1) & (lSel == 1)) {
 
     # If we're working with matrices
     if (is.matrix(data)) {
@@ -186,16 +201,15 @@ select <- function(data, selection, cols = TRUE) {
                       dimnames = list(rownames(data), colnames(data)[selection]))
 
       }
+      # Create a matrix of 1 row
       else {
 
-        # Create a new matrix of one row
         out <- matrix(data[selection,], nrow = 1,
                       dimnames = list(rownames(data)[selection], colnames(data)))
-
       }
 
-    } # Working with matrices
 
+    } # Working with matrices
 
     # If we're working with a data frame
     else {
@@ -213,17 +227,94 @@ select <- function(data, selection, cols = TRUE) {
       else {
 
         # Create a new data frame with one row
-        out <- data[selection,]
+        out <- data.frame(data[selection,], row.names = rownames(data)[selection])
+        colnames(out) <- colnames(data)
 
       }
 
     } # Working with data frames
 
-  } # Selecting a single row or column
+      
+  } # Case 2
+
+  # Case 3:  n x 1, select 1 row, select multiple rows, select 1 column
+  else if ((nRow > 1) & (nCol == 1)) {
+
+    # Selecting 1 column
+    if (cols) {
+      out <- data
+    }
+
+    # Selecting 1 or multiple rows
+    else {
+
+      # Matrixes
+      if (is.matrix(data)) {
+  
+        out <- matrix(data[selection, 1], ncol = 1,
+                      dimnames = list(rownames(data)[selection], colnames(data)))
+      }
+      
+      # Data frames
+      else {
+  
+        isChar <- is.character(data[,1])
+        out <- data.frame(data[selection, 1], row.names = rownames(data)[selection], stringsAsFactors = !isChar)
+        colnames(out) <- colnames(data)
+        
+      }
+        
+    }
+      
+  } # Case 3
+
+  # Case 4:  1 x n, select 1 row, select multiple columns, select 1 column
+  else if ((nRow == 1) & (nCol > 1)) {
+
+    # Selecting a single row
+    if (!cols) {
+      out <- data
+    }
+
+    # Selecting 1 or more columns
+    else {
+
+      # For a matrix
+      if (is.matrix(data)) {
+
+        out <- matrix(data[1, selection], nrow = 1,
+                      dimnames = list(rownames(data), colnames(data)[selection]))
+            
+      }
+
+      # For a data frame
+      else {
+
+        out <- data.frame(data[1, selection], row.names = rownames(data))
+        colnames(out) <- colnames(data)[selection]
+
+        # If only 1 column was selected, the string will have been converted to a factor
+        if ((lSel == 1) & (is.character(data[1, selection]))) {
+          out[,1] <- as.character(out[,1])
+        }
+        
+      }
+
+    }
+
+  } # Case 4
+
+  # Case 5:  1 x 1
+  else if ((nRow == 1) & (nCol == 1)) {
+     out <- data
+  }
+
+  # No more cases should remain
+  else {
+    stop("Unexpected error.  This should not have happened")
+  }
 
   # Return the result
   return(out)
 
 } # select
-
-